requirements.txt
boto3==1.26.74
pytest==7.2.0
python-dotenv==0.21.0
requests==2.31.0
Flask==2.2.2
waitress==2.1.2
flask_cors==3.0.10
pybase64==1.3.0
autopep8==2.0.4
pylint==3.0.2
Werkzeug==2.2.2
openai
langchain
tiktoken